feat: 新增sdk公共模块 #6867
