+++ src/backend/ci/core/repository/biz-repository/src/main/kotlin/com/tencent/devops/repository/resources/ExternalRepoResourceImpl.kt
@@ -30,16 +30,29 @@ package com.tencent.devops.repository.resources
 import com.tencent.devops.common.web.RestResource
 import com.tencent.devops.repository.api.ExternalRepoResource
 import com.tencent.devops.repository.service.scm.IGitOauthService
+import com.tencent.devops.repository.tapd.service.ITapdOauthService
 import org.springframework.beans.factory.annotation.Autowired
 import javax.ws.rs.core.Response
 import javax.ws.rs.core.UriBuilder
 
 @RestResource
 class ExternalRepoResourceImpl @Autowired constructor(
-    private val gitOauthService: IGitOauthService
+    private val gitOauthService: IGitOauthService,
+    private val tapdService: ITapdOauthService
 ) : ExternalRepoResource {
     override fun gitCallback(code: String, state: String): Response {
         val gitOauthCallback = gitOauthService.gitCallback(code, state)
         return Response.temporaryRedirect(UriBuilder.fromUri(gitOauthCallback.redirectUrl).build()).build()
+    }
+
+    override fun tapdCallback(code: String, state: String, resource: String): Response {
+        val uri = UriBuilder.fromUri(
+            tapdService.callbackUrl(
+                code = code,
+                state = state,
+                resource = resource
+            )
+        ).build()
+        return Response.temporaryRedirect(uri).build()
     }
 }
