+++ src/backend/ci/core/repository/plugin-tapd/src/main/kotlin/com/tencent/devops/repository/tapd/config/TapdProperties.kt
@@ -1,7 +1,7 @@
 /*
  * Tencent is pleased to support the open source community by making BK-CI 蓝鲸持续集成平台 available.
  *
- * Copyright (C)) 2019 THL A29 Limited, a Tencent company.  All rights reserved.
+ * Copyright (C) 2019 THL A29 Limited, a Tencent company.  All rights reserved.
  *
  * BK-CI 蓝鲸持续集成平台 is licensed under the MIT license.
  *
@@ -11,7 +11,7 @@
  * Terms of the MIT License:
  * ---------------------------------------------------
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
- * documentation files (the "Software")), to deal in the Software without restriction, including without limitation the
+ * documentation files (the "Software"), to deal in the Software without restriction, including without limitation the
  * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
  * permit persons to whom the Software is furnished to do so, subject to the following conditions:
  *
@@ -25,20 +25,18 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-dependencies {
-    api(project(":core:common:common-service"))
-    api(project(":core:common:common-web"))
-    api(project(":core:common:common-client"))
-    api(project(":core:common:common-auth:common-auth-api"))
-    api(project(":core:ticket:api-ticket"))
-    api(project(":core:process:api-process"))
-    api(project(":core:common:common-scm"))
-    api(project(":core:repository:api-repository"))
-    api(project(":core:repository:model-repository"))
-    api(project(":core:repository:plugin-tapd"))
-    api(project(":core:common:common-db"))
-    api(project(":core:auth:api-auth"))
+package com.tencent.devops.repository.tapd.config
 
-    api("com.google.code.gson:gson")
-    api("org.eclipse.jgit:org.eclipse.jgit")
-}
+import org.springframework.boot.context.properties.ConfigurationProperties
+
+@ConfigurationProperties(prefix = "tapd")
+data class TapdProperties(
+    val serverUrl: String,
+    val apiUrl: String,
+    val clientId: String,
+    val clientSecret: String,
+    // tapd回调地址
+    val callbackUrl: String,
+    // tapd授权后跳转页面
+    val redirectUrl: String
+)
